move dmsg-server config from cmd to pkg
